requirements.txt
fastapi
uvicorn
gitingest
lightrag-hku
requests
nest_asyncio
aiofiles
ascii_colors
asyncpg
distro
graspologic>=3.4.1
httpcore
httpx
jiter
numpy
openai
passlib[bcrypt]
pipmaster
PyJWT
python-dotenv
python-jose[cryptography]
python-multipart
pytz


aiohttp
configparser
future

# Basic modules
gensim

# Additional Packages for export Functionality
pandas>=2.0.0

# Extra libraries are installed when needed using pipmaster
pydantic

setuptools
tenacity

# LLM packages
tiktoken
xlsxwriter>=3.1.0